Fix reference error for class literals in SDM.
If during the initial SDM compile a class literal was not referenced,
the class literal would not be part of the JS for that class. Instead
it would be generated in the epilogue in the compile were a reference
is introduced.

The epilogue does NOT have memory, so if a subsequent recompile did
not see a reference to the mentioned class literal it would not
be part of the epilogue and a reference error would occur.

Given that UnifyAST did not prune anything and class literal references
are synthezised at each class getClass(), the only class literals for
interfaces could cause the error (making this error a rare occurrence).

When compiling without optimizations (a requirement for SDM), this
patch creates class literals for all types in the current compile.

Change-Id: Ie5f287794bf66a2988ae07f4fda0cd6468c030f4