Update patch set 4
Patch Set 4:

I am giving up on this patch. There is 50% performance regression in iteration of the map.

I filed an issue with V8 [1] but there is no clear timeline on when this will be improved.

I will create another patch that utilizes the cleaner design here with the Object.create based implementation. This patch can be rebased on top of that when the performance catches up.

[1] https://code.google.com/p/v8/issues/detail?id=4014

Patch-set: 4