Update patch set 2
Patch Set 2: Code-Review+2

> Hi Thomas. Can you +2 if you're ok with the change? Build is still
 > broken.

I was actually waiting for some additional feedback weighing for one proposal or the other.

While I think it might actually probably be best to just "implement" the methods in the invoke() method instead of "routing" them to the InvocationHandler instance, this is the approach used in other InvocationHandlers so if only for consistently I'm going to go ahead and merge this.

@Andreas, could you abandon your change then?

Patch-set: 2
Label: Code-Review=+2