Updates jsinterop the check in MakesCallsStatic.
Earlier check was suspectable as it was only performed for JsType
methods but not JsFunction methods. Then I noticed that I cannot
reproduce the problem but removing the check was breaking the
compiler and causing an exception while performing inlining of
super calls to jstype prototypes. This is because those super
calls were incorrectly being tried to statified.
Patch replaces the special case so calls to JsType prototypes are
no longer statified.

Change-Id: Ic0dfb79d7b425b5961a2bbb85f542df95fbbd920