Update patch set 5
Patch Set 5:

The LogDisabled thing only would have had an effect in SDM (otherwise the setup was broken anyway).

Here is what it did:
When running in SDM it added the extra SimpleConsoleLogHandler (I added a hardcoded line for that in LogManager).

So you might end up having the console handler double in your SDM setup which I think is fine, WDTY?

Patch-set: 5
Label: Verified=0