Update patch set 6
Patch Set 6:

> How about we add another config property from the Logging.gwt.xml
 > and detect that and if its set do not add the console logger?

Sounds like a good idea but in GWT (and maybe for J2CL in future), the compilation fails if the property is not defined (we can workaround it but that will be ugly). I think we can modify your idea to make it work.

Let's first define the behavior we want. Thinking out loud here;
1- For a new app, if logging is enabled, we should add the simple console logging by default.
2- If running in dev mode, we should enable logging by default.
3- For old apps that inherit complex logging, we should disable simple console handler.

So let's add gwt.logging.simpleConsoleHanderEnabled. As default value in System.getProperty we can pass value of super-dev-mode. Then we can set it to false from LoggingDisabled.gwt.xml. That should give us the behavior defined above.

Patch-set: 6
Label: Verified=0