Fixes instanceof behavior for concrete JsTypes.
Earlier removed hasLiveImplementors check has a side
effect that was rejecting non-interface types. As a
result this was causing correct instanceof check to
be chosen. Now doing it properly...

I am still not sure how it was working correctly with
hasLiveImplementors check in place (as that should cause
regular java checking if no live implementor exist).
Anyway I am not motivated to debug that now...

Change-Id: Ib9b847935c7922e6d82bf2e9333aed8df04a843b