Refactor the compiler around new JsInterop semantics.
One of the singificant change is being a jstype no longer
pins instance members. Only thing that pins any member is
being exported or not. However this shouldn't reflect to
end user as the current JsType annotation automatically
marks the instance members as exported to keep the existing
behavior intact. The new annotations will NOT automatically
mark them exported so they will not be pinned for optimization
purposes.

Change-Id: I5ab76d42f295bab0937f9b88d527d7859a5da055