Speed up sdm - cache idempotent class resolutions
This modification reduces in average by three times the
first compilation during the SuperDevMode starting
process.

Fqn are resolved along the persing of the Java files
by JDT. During the process the same fqns are processed
again and again, wasting time and processing power.

By applying a caching strategy, the time for the
startup compilation of the super dev mode is
divided by three and sometimes more depending on
the project.

Change-Id: I1b0ea3798b5d3afc15cfbe5731e7eca2a9623b7f