Update patch set 2
Patch Set 2:

> Any ideas on the best ways to update EqualityNormalizer according
 > to my commit log comment? The 2D matrix approach looks a little
 > ugly, I wonder if I should just reimplement it in imperative logic
 > without a lookup table.

I think that it is already incorrect for Strings ("" == null) in Java gets transformed to (!!"") which is not correct. 

We might reuse this and instead of STRING/NOTSTRING would be JSPRIMITIVE/NOTJSPRIMITIVE and the comparisons JSP/JSP, JSP/NULL and UNK/NULL an should be 1 instead of 0.

Patch-set: 2