Update patch set 1
Patch Set 1:

> (1 comment)

I'm fine with leaving it as a warning.  I don't want to overcomplicate the patch given the performance improvements it brings and it simplifies other stuff.

The '+' operator stuff could be a heisenbug though, imagine a POJO field that's a long is updated and ends up a string, and then it's posted to the server where there's an attempt to insert it into a DB.  It'll be real hard for someone to track down exactly where the concat is occurring that ended up serializing a bogus String to the server.

Patch-set: 1