Update patch set 1
Patch Set 1:

> > (1 comment)
 > 
 > I'm fine with leaving it as a warning.  I don't want to
 > overcomplicate the patch given the performance improvements it
 > brings and it simplifies other stuff.
 > 
 > The '+' operator stuff could be a heisenbug though, imagine a POJO
 > field that's a long is updated and ends up a string, and then it's
 > posted to the server where there's an attempt to insert it into a
 > DB.  It'll be real hard for someone to track down exactly where the
 > concat is occurring that ended up serializing a bogus String to the
 > server.

I'm not planning to introduce a warning in this CL. @leafwang already doing something similar with JsOpaque so we can handle it there. (probably we can replace whole JsOpaque with a warning and treat jsinterop warning as errors by default then use SuppressWarnings instead to silence them; so people will not need to learn a new annotation).

Patch-set: 1
Label: Verified=0