Fix string concatentation when both strings are null.
String concatenation is translated directly into + in
JavaScript which introduces a semantic difference.
Given string varibles a and b, a + b in JavaScript is 0
when both a and b are null. This patch appends a + "" to
binary operations where both the left hand side and the
right hand side can be null.

Bug-Link: https://github.com/gwtproject/gwt/issues/8247
Change-Id: Ice4b39cab9efd8310fc97c11821918976747b44f