Update patch set 6
Patch Set 6:

Discussed offline and decided to go this route to make the devirtualization more explicit by having a separate class and doing this by hand instead of compiler devirtualization.

Let's move it as a package visible class under java.lang and rename it to something more explicit; e.g. DevirtualizedString.

Don't forget to do the Boolean and Double as well.

Patch-set: 6