Update patch set 1
Patch Set 1:

I actually described in the commit msg why it is not com.google.gwt or org.gwtproject.
This annotation are not specific to GWT-SDK. Similarly, that's why we chose javaemul.internal instead of com.google.gwt.core earlier.

I put them to a different source root because it should be easy to consume without any other gwt dependency (Internally they are a different build target and available as a separate jar). Putting them in same source root with other and trying to extract from there seems less desirable to me but I'm open to other suggestions.

Patch-set: 1