Update patch set 7
Patch Set 7:

(1 comment)

To me this still seems very awkward and requires a lot unnecessary copy paste around. When we add a field to one class, we need to copy paste to another class. When we add a method. we need to do the same. To catch any errors, we will even need to write tests for simple static fields, constants etc. So much waste!

On the other side, the only reason of making this to make the devirtualization more explicit but everything becomes more complex when you starting thinking about static field, static method, clinit etc. I don't think this is simpler mental model.

Looking at this more I'm getting more convinced on the "prototype" approach that I described earlier.
In that approach we basically keep the classes simple, intact and generate instance calls these classes as String.protoype.x.apply , Double.prototype.x.apply etc. Very simple to implement, very simple maintain, very simple to understand. We didn't go with this one before because the earlier implementation of the Double was not compatible but it is now.

Patch-set: 7