Update patch set 19
Patch Set 19:

(14 comments)

First pass done. Just a general comment, this is a big change, but you also cleaned up a lot of other stuff in unrelated classes along the way that had not much to do with GenJsAST, it would probably be better in the future to keep those in separate small CLs.

Patch-set: 19