Implement JsOpaque check.
The purpose is to prevent from passing unmatched variables
at JS side accidentally.

For example,
@JsType
interface MyJsInterface {
  void foo(A a);
}

If A doesn't have any accessible methods from javascript (
not one of the following types:
JsType, JSO, JsFunction, non-long primitive types, Boolean,
Double, String, Object, return type of any exported methods
) This is a compile error.

But if this is intentional as the parameter will be only treated
as an opaque object (e.g. a Map), the parameter or the method
should be annotated with @SuppressWarnings("JsOpaque").

Change-Id: Ic2fe9efb520d6350716add31d0172f455c524791