Fixes SDMv to incl. implementors of interfaces in stale types
Earlier implementors of interfaces were not included in
subtypes hence when an interface was changed, the callers
of implementors were not considered stale. However they
need to be consider stale and recompiled (due to defender
methods, JsInterop, etc.).

The patch also converts last piece of CompilerTest to run
with new jsinterop annotations.

Change-Id: I75296d7ee10251b2968a522547243a61d3a71bdc