Update patch set 1
Patch Set 1:

This is the pattern I'm suggesting to follow for Java8 test exclusion (this is necessary because we use Java7 JVM and should not try to compile the test in java).
An exclusion similar to here should be added to ant files however I'm ant illiterate so you guys need to update it.

Patch-set: 1
Label: Verified=0