Fix resource scanning for Windows file conventions.
GWT internally uses "/" as a path separator, the old resource code
used to construct paths one directory at a time (so it would never
have to process file separators). The new resource code recieves
paths from the operation system, hence there is a need to normalize
them for the rest of the code.

Change-Id: I1794314224a55b6fa7a2c1f1e9084b10da536444