Update patch set 3
Patch Set 3:

Agreed a test is preferable; I don't see a UiBinderParserTest, but there is UiBinderParserUiWithTest...this could probably be used as a basis for UiBinderParserUiImportTest.

Mark, does that look feasible?

I'm glad Thomas did not -1, as I agree about a test, but I don't want it to hold up a valid fix.

Patch-set: 3