Update patch set 7
Patch Set 7:

well, you're missing part of the "gwtrandombug" fix. Besides the change you here, the SerializableTypeOracleBuilder in that zip file also has this change:

@@ -1047,8 +1049,12 @@ public class SerializableTypeOracleBuilder {
     JClassType[] allTypes = typeOracle.getJavaLangObject().getSubtypes();
     for (JClassType cls : allTypes) {
       if (isDeclaredSerializable(cls)) {
-        computeTypeInstantiability(localLogger, cls, TypePaths.createSubtypePath(parent, cls,
-            typeOracle.getJavaLangObject()), problems);
+        // FIX so that the order of classes is not important.
+        TypeInfoComputed tic = ensureTypeInfoComputed(cls, null);
+        if (tic == null | !tic.isPendingInstantiable()) {
+          computeTypeInstantiability(localLogger, cls, TypePaths.createSubtypePath(parent, cls,
+              typeOracle.getJavaLangObject()), problems);
+        }
       }
     }
   }
@@ -167

It does not explain why your test hasn't reproduced it so far though. My current assumption is that the gwtrandombug compile is invoking more of the gwt compile cycle than just the "build type oracle" that this unit test is doing.

Patch-set: 7