Update patch set 7
Patch Set 7:

Per the unit test, I put this at the start of computeTypeInstantiability:

    System.out.println("computeTypeInstantiability " + type + " ... " + path);

And when I run the unit test, there are ~10 lines of output.

When I build a gwt-user.jar, and put it though the gwtrandombug reproduction, it outputs a similar ~10 lines of output, but then a whole bunch more. I haven't poked around at why yet, but that is what I see so far.

Patch-set: 7