Fix implemention of Object methods invoked a native object.
Object methods on native objects are handled by the JSO trampoline.
For that to work corretly, we make sure that we never tighten a
method call into a native JsMethod. There are no performance
benefits of doing so, and it interferes with the hacky way we
are handling these overrides.

Change-Id: I522cf60ae6a0077c0c009fc4e081c75e5c85acb1