Update patch set 1
Patch Set 1:

> Is there a good reason using a google'ish "goog" prefix for these
 > property names or was it just habit? IMHO it does not add value and
 > just makes property names longer. Also feels like these properties
 > have a tight coupling to Google. I assume its somehow related to
 > J2CL but then it would be nice to have a bit more context in the
 > commit message, so non googlers can follow as well.
 > Currently these property names will simply look unusual compared to
 > all other GWT properties.

@Jens: Yes, I should have put better description.
As Daniel said, this is in preparation for the long run for better closure integration and trying to be more friendly the way closure deals with properties (which are globals and prefixed by project where the core is named 'goog').
If you like, you can you sen a patch that preserves "checks.*" and  forwards to new properties. That seems like a good idea to me. Let me know if you are planing to do that.

Patch-set: 1