Make exceptions behave well across Java/JavaScript boundary.
The essence of the change is, instead of throwing subclasses of Java
Throwable class in generated throw statements, we now throw the backing
JavaScript error object.

As the propagated thrown object is a real JavaScript error, it plays
better with both JavaScript callers (which is essential  via JsInterop)
and also the browser dev tools.

There are two potentially breaking semantic changes with the patch:

 - If the JavaScript original throw encapsulated by JavaScriptException
is not a JavaScript Error (e.g. a string as you can do that in JavaScript),
we no longer auto-fill the stack-trace by other means. The developers needs
to call fillInStackTrace in that case. I think this we was already the behavior
pre GWT 2.6 or 2.7 (see the original javadoc that wasn't fixed) and preserving
the newer behavior in the new design causes extra complications.

 - If you catch the exception in JSNI and pass it back to Java, it is not a
Java object anymore. This is probably not harmful for anyone.



Change-Id: Id21c182078075ca19bb0953136a8b9136adcba63