Update patch set 2
Patch Set 2: Code-Review+2

The patch is ok, and works in my local.
We have to synchronise merging and changing the build script in jenkins.
So let me merge this once I do a couple of tests of the patch in the CI server.
I think is ok to have the addition to .gwt.xml here

Patch-set: 2