Make exceptions work across Java/JavaScript boundary
The essence of the change is, instead of throwing subclasses of Java
Throwable class in generated throw statements, we now throw the backing
JavaScript error object so the generated code looks like this:

 try {
   throw new SomeJavaException.backingJsObject; // Exceptions.unwrap
 } catch(e) {
   e = Exceptions.wrap(e);

   if (instanceof(e, SomeJavaException_ID)) {
     // catch block for SomeJavaException
   }
   // other catch blocks...
   else {
    throw e.backingJsObject; // Exceptions.unwrap
 }

As the propagated thrown object is a real JavaScript error, it plays
better with the browser dev tools and JavaScript callers (which become
more essential due to JsInterop).

There are two potentially breaking semantic changes with the patch:

 - If the original thrown object encapsulated by JavaScriptException
is not a JavaScript Error (e.g. a string as you can do that in JavaScript),
we no longer try to auto-fill the stacktrace by other means. The developers
needs to call fillInStackTrace in that case. I think this we was already the
behavior pre GWT 2.6 or 2.7 (see the original javadoc that was missed) and
preserving the newer behavior in the new design causes extra complications.

 - If you catch the exception in JSNI and pass it back to Java, it is not a
Java object anymore. This is unlikely to be a real issue in practice.

The patch is also breaks GIN as it uses GIN has its own hacky class loader
to load super-sourced versions of the files and ends up trying to load java.lang.JsException which is forbidden by java due to package name.
I fixed GIN but this could be also worked around in released versions by
setting a compiler flag:
 -setProperty gin.classloading.exceptedPackages=com.google.gwt.core.client,
com.google.gwt.core.client.impl

Change-Id: Id21c182078075ca19bb0953136a8b9136adcba63
Review-Link: https://gwt-review.googlesource.com/#/c/14030/