Update patch set 2
Patch Set 2:

w.r.t to "eval", although is valid to define variables names like "eval", "arguments" or even "undefined", we don't want to be generating obfuscted (or pretty) names that clash with this.

Bottom line, it is always safe to add to the list of illegalnames.

Patch-set: 2
Label: Verified=0