Update patch set 1
Patch Set 1:

> @rluble: Our original idea was to remove all kind of casts for the
 > base Object class so it could be assigned from anything, wasn't it?

Yes. I am not sure what is easier to communicate. We need to implement the isInstance behaviour so that casts/instanceof can be customized. The question is what makes sense as a default and I think that if we are doing Array.isArray for GLOBAL.Array it might be sensible to do typeof == "Object" for GLOBAL.Object. Maybe we need to introduce JsName.ANY for untyped things (or maybe "*"). These JS quirks leave no elegant option on the table.

Patch-set: 1