Update patch set 5
Patch Set 5:

> I uploaded an alternative for the NpeTest and reverted others. Does
 > it look good to you?

I am not sure how you expect this to work with --nogenerateJsInteropExports. backingJsObject is defined as a JsProperty in class Throwable. Then it is accessed in Java as well as in JSNI with the exported name. When --nogenerateJsInteropExports is selected there is no guarantee that the JsName for that property is preserved (and it is not)

Patch-set: 5