Add treatment for special native types.
Native JsTypes can model the some of the standard builtin JavaScript objects like
Object, Function, Number, Array and String. Due to the quirks of JavaScript creating
this objects in the current IFrame or in the window scope results in different
constructors, messing with casting and instance. This patch makes the aforementioned
JavaScript "classes" behave in a more coherent manner.

Change-Id: I111f2007f7d67968c74dbd8cb92321f4ddf24b27