Make symbolMaps for StackTraceDeobfuscator lookup more robust
Users can specify the symbolMaps location for the RequestFactoryServlet
by adding an init-param  symbolMapsDirectory to the web.xml. However
the RequestFactoryServlet only passes the user specified directory to the
Logging class. This does not work in WAR deployment scenarios.

To solve this, the RequestFactoryServlet will check if a
symbolMapsResource init parameter is defined and pass it to
ServletContext#getResource to retrieve the correct URL to the
symbolMaps directory.

 Bug: #6363
 Bug-Link: https://github.com/gwtproject/gwt/issues/6363

 Change-Id: I916b687ab59b703f8aee85bd7ea1e96f828c0186

Change-Id: I0ec628b2e4a7401e2b076f07bd22dc8e09972c7c