Update patch set 1
Patch Set 1:

We we want to pull ahead with this we need some kind of way of saying is this object created in java or javascript.

Right now in some places (where people used to rely on isntanceof JSO) I added obj.getClass().equals(JSO.class) which will break with this change.

I am patching in and running a global presubmit to see if this is even feasiable

Patch-set: 1
Reviewer: Gerrit User 5385 <5385@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=0