Update patch set 2
Patch Set 2:

The original behavior was "instanceof Object" and that was consistent with the rest of the spec.
Later we changed that to 'typeof 'Object'" to make it work across the frame;
which all sounds reasonable to me.
So I don't understand why the semantics are confusing?

Patch-set: 2
Reviewer: Gerrit User 5110 <5110@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=0