Use canonical IP address when bindAddress is 0.0.0.0
When a wildcard bindAddress is specified, make a guess
as to the best IP to use by looking up the IP for the canonical
host name.  Always use the IP address to support
environments where a proper DNS entry is not available.

Continue to use the bindAddress as entered for everything else.

This will fix SDM source mapping issues with: -bindAddress=0.0.0.0

Closes #8910 on GitHub
Closes #8467 on GitHub

Change-Id: Ifd0a819e3169c1635388acf6b5c4cfa51e5f2f1c