Fix SDM source mapping issues with -bindAddress=0.0.0.0
When a wildcard bindAddress is specified, make a guess
as to the best IP to use by looking up the IP for the canonical
host name.  Always use the IP address to support
environments where a proper DNS entry is not available.

Continue to use the bindAddress as entered for everything else.

https://github.com/gwtproject/gwt/issues/8910
https://github.com/gwtproject/gwt/issues/8467

Change-Id: Ifd0a819e3169c1635388acf6b5c4cfa51e5f2f1c