Update patch set 33
Patch Set 33:

> > Can someone schedule a verification build on: [1], please? Could
 > I
 > > do it myself?
 > >
 > 
 > I run the build, but the test continues failing:  http://build.gwtproject.org/job/gwt.presubmit.runtests/9/testReport/

Yes, looking into it. It seems to be related to newer JS package htmlunit-core-js-2.17.jar (LogRecord in RPC was already missing the obfuscated stack trace, because it wasn't correcty emitted on the client side). So instead of disabling yet another test, I went for downgrade the upgrade to htmlunit-core-js-2.15.jar instead. This and another native tests, that i disabled in previous patch set are passing now. Clearly, the verification build is failing now because i would need to upload another PR to GWT_TOOLS GH repository with older JS core package. I'm working on it right now and hope that someone can approve the PR.

Patch-set: 33
Label: Verified=0