Update patch set 33
Patch Set 33:

>  Can this change be approved now?

Well, that is a decision of the SC if they want this in 2.8.0.
Since 2.8.0 is a breaking release because needs at least java7 and it's thought to bring modernity to GWT, IIMHO upgrading old jetty and htmlunit should be a nice feature in 2.8.0 as well.

On my part I have deeply tested this patch, I'm using it in my prototyping projects for a long and it works like a charm. With it I can run vaadin projects with the devmode something that it was pretty difficult before because of class collisions.

The only thing that I miss, is that the launcher is not configured to parse servlet 3 annotations (when jetty annotations is in the classpath),  as well as it does not support hot reloading when server sources changes. But it might be discussed in a separate patch.

Patch-set: 33
Label: Verified=0