Update patch set 33
Patch Set 33:

> I must say I'd have rather separated the upgrade to a newer Jetty
 > and the "unbundling" of Jetty from gwt-dev.

OK. I don't have problem with this. If this change is going to be included in 2.8 release, then it's clearly an advantage to have not that intrusive change, that affects final artifacts. So I will remove unbundling Jetty deps from this change and upload amended version later today.

Patch-set: 33