Update patch set 1
Patch Set 1:

> I feel like ArgHandlerSourceLevel::setString should return always
 > true or at least might not fail with 1.7, so as maven projects
 > using the maven.compiler.source continue working. Notice that
 > gwt-maven-plugin is reusing this property for setting the gwt
 > source level, probably it should use a different one.

I think the breakage is a good thing actually.
When updating GWT, you'd also have to set `sourceLevel` to 8 unless you want to update everything to Java 8.
We're assuming Java 8 is backwards-compatible with Java 7 (at source level), which is probably true, but I'd bet there are a few differences (and/or bugs in JDT), so it's a good thing IMO to be explicit iff you're not using the same "source level" for JavaC and for GWT.

 > Another option is remove completely the  sourceLevel argument,
 > since gwt compiler always consider 1.8

That would break everyone, more badly than the above (Mojo's plugin *always* passes sourceLevel it'd break all users until a new version of the plugin is released; and my plugin passes sourceLevel if not null, and because it's initialized with a property, that would mean updating POMs in possibly non-straightforward ways)

An alternative would be to accept 1.7 as a "legacy" value, and turn it into 1.8 but *not* silently (emit a warning) *and* continue to reject all other values.

Patch-set: 1
Reviewer: Gerrit User 5030 <5030@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0