Update patch set 7
Patch Set 7:

> It seems that this patch breaks almost uibinder tests.
 > http://build.gwtproject.org/job/gwt/lastCompletedBuild/testReport/

I just looked into it. I thought the syntax in this CL was correct but turns out there are three issues:

  - The import must be either "com.google.gwt.uibinder.test.client.Constants.InterfaceConst.*" or "com.google.gwt.uibinder.test.client.Constants.InterfaceConst.someValue"
  - The ui:text reference must then be "{someValue}" 
  - The <ui:import> does not work when put inside a widget element. Only seem to work when its at the top level of the uiBinder tag either at the beginning or at the end.

Turns out when I fix the three things above all UiBinder tests work again including the one added in this CL. Unfortunately the test added in this CL also works without the change to UiBinderParser. So the added test is kind of pointless to verify if the change to UiBinderParser fixes the references issue.

Should we just revert the whole commit or should I push a follow up which only removes the added test but keeps the UiBinderParser change (I believe the code is more correct than before)?

Patch-set: 7