Fix AutoBeanFactoryGenerator in preparation for Spliterator/Stream addition
Three bugs are fixed here, to support new Java8 methods which will be
added soon to GWT. These are written to be general enough to support
these same techniques in user-created autobeans.

First, don't attempt to shim static methods - they aren't needed, and
they confuse the compiler.

Second, when ModelUtils is asked for a base type, it should know that
isWildcard can return type parameters, so unwrap again.

Third, since generated autobean instances already throw away generics,
cast all return values to the type that we expect them to be, since the
compiler will not be able to figure it out on its own due to other raw
generics.

Change-Id: I605d23ff2c1ed29f5ff512290c22d7f2b73bf23c