Update patch set 1
Patch Set 1: Code-Review+1

> (3 comments)

Is allows to allocate underlying array of exact size and prevent data copying if the size of argumetn in addAll() is > 10. So the operation of creating new ArrayList and its population with data is perfromed faster and memory is allocated only once at creation time and of exact size.

 > (3 comments)

Patch-set: 1
Label: Code-Review=+1