Update patch set 1
Patch Set 1:

What's the use-case? The XMLHttpRequest will be passed to the Callback upon request completion, so the only use-case I foresee is to be able to abort() the request. It'd then be possible to return a small/simple object only exposing that abort() method, similar to com.google.gwt.http.client.Request.

Patch-set: 1
Label: Verified=0