Update patch set 2
Patch Set 2: Code-Review+1

Would you mind adding a test to GwtSpecificValidatorCreatorTest? There's a testAsLiteral_1L that tests that the 'L' suffix is there, but there would need an additional test that checks that the value is not truncated (e.g. using 9999999999L as input like in the bug report)

Patch-set: 2
Label: Code-Review=+1