Remove accidental override removal optimization.
Accidental override removal is currently implemented in
GenerateJavaScriptAST and not reflected in the AST. In
particular causes CodeSplitter to not account for the
removal and incorrectly place its super method.

This patch limits the optimization to a case that is safe
and necessary. The case in question is an accidental
override that delegates to a JsMember; and its removal is
needed because the compiler does not handle dispatch to
JsProperties.

An optimization to remove all unnecesary accidental overrides
is possible but needs to be implemented a Java AST
transformation.

Change-Id: I41593d5fc951d3afe705a5c8472ca4252785a424