Tentatively ignore temp files used for "safe writes" while watching files
Some IDEs (e.g. IDEA with "safe writes" turned on) will write to a
temporary file before moving to replace the actual file; this caused the
ResourceAccumulator to see an ENTRY_CREATE event but the file is deleted
before it could be looked at, and a NoSuchFileException is thrown as a
consequence.

Bug: #9310
Bug-Link: https://github.com/gwtproject/gwt/issues/9310
Change-Id: I3567bcc3ef899cd807f2ee580069992cc87a814b