Update patch set 6
Patch Set 6:

Michael: Have you actually tried using the Closure compiler back end? It is activated with -XenableClosureCompiler.

There is a failing check if you enable it:
   [ERROR] Unexpected internal compiler error
java.lang.IllegalStateException: Externs zip must match our hard-coded list of externs.
	at com.google.gwt.thirdparty.guava.common.base.Preconditions.checkState(Preconditions.java:173)
	at com.google.gwt.dev.js.ClosureJsRunner.getDefaultExterns(ClosureJsRunner.java:123)
	at com.google.gwt.thirdparty.guava.common.base.Preconditions.checkState(Preconditions.java:173)
	at com.google.gwt.dev.js.ClosureJsRunner.getDefaultExterns(ClosureJsRunner.java:123)

That seems to be a sanity check that we are using the closure compiler version that we are expecting. 

Can you fix that? Otherwise I'll revert this cl.

Patch-set: 6