Update patch set 2
Patch Set 2:

Just to keep in mind; lambda are not innocent as they look in the code for GWT.
It is not like a passing a function; we basically synthesize a class, instantiate and do method dispatching; so they will add significant code bloat and slowness. So until we have a better optimized code path for them, try to avoid temptation to use it.

Patch-set: 2
Reviewer: Gerrit User 5110 <5110@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Verified=0