Update patch set 4
Patch Set 4: Code-Review+2

There is no need to do this type of changes now; we are in the process of releasing 2.8 and we should avoid doing unnecessary changes. 

We normally tackle deprecation of things we depend on when we upgrade the dependencies.

Patch-set: 4
Reviewer: Gerrit User 5105 <5105@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=+2