Place class literal in a correct fragment for -XenableClosureFormat.
When -XenbleClosureFormat is specified the code that links class
literals to the class variable is emitted with the class definition.
This situation generates a new load order dependecy.

This patch makes sure that class literal creation happens before
class definition, and also makes sure that if a enum class literal
is moved then also are the corresponding values() and valueOf()
methods.

Change-Id: I1cb32ce70dd37e4c3a7b0e6a933e529dd0347302