Update patch set 3
Patch Set 3:

(2 comments)

> (2 comments)
 > 
 > How about excluding the json/impl/Jre* files from client code,
 > given that it's clearly labelled as being a server-side
 > implementation?

Yes that is the first I tried adding a rule in gwt.xml, but actually client implementation use the same classes, so we need some extra work.
What could be wrong is naming these classes with the Jre suffix. Well they are implementations using pure java which works in JVM and emulated Jre so it could be the reason for their names.

Patch-set: 3