Update patch set 6
Patch Set 6:

It appears that this patch has broken a bunch of tests. When running EmulJava8Suite, I get several failures that were not present before this change (all in PrimitiveIteratorTest or StringJoinerTest):

https://gist.github.com/niloc132/2b4f2a09de6508fe3bc4194842a2cb5f

Command I am using to get this result:

$ ant test.web.htmlunit -Djavac.source=1.8 -Djavac.target=1.8 -Dgwt.junit.testcase.includes=**/EmulJava8Suite.* 

Propose we revert this, or a new commit needs to permit these checks to *not* throw as is currently expected.

Patch-set: 6