Update patch set 19
Patch Set 19:

(1 comment)

My mistake: 9333 has workarounds in this (all already have a TODO), but 9340 only applies to the stream patch.

The suggestion has been made to add InternalPreconditions calls to this - I'd propose that this happen in a follow-up patch to improve errors and not prevent this from going out in the upcoming beta/rc. I'll start on it now base don Jens's initial suggestions.

Patch-set: 19
Label: Verified=0