Update patch set 3
Patch Set 3: Code-Review+2

com.google.gwt.dev.Compiler API probably was never meant to be public, but this is a breaking change. Are we OK with it?

(I believe we are, but checking just in case)

Patch-set: 3
Reviewer: Gerrit User 5030 <5030@84fdd5b9-d3fc-3149-9ae9-e619898c4c46>
Label: Code-Review=+2