Update patch set 2
Patch Set 2: Code-Review+2

The only technical concern is that exposing default and static methods through an existing API might break existing generators that are not expecting these. And given the prevalence of static and default methods in the JRE now it might require fixing some of the existing generators.

The other alternative is to hide this methods from the existing API and possibly exposing them through an alternative API.

That said, I'll leave the decision to you guys.

Patch-set: 2