Update patch set 2
Patch Set 2:

The flip side is for external generators that correctly emit @Override on a method they are overriding, but with a static method that must be overriden according to the typeoracle. It isnt an abstract method, but we're claiming it is, and compilation will fail when it sees @Override in Java7+.

Or are we comfortable with forbidding any generator from implementing methods on interfaces which happen to decide to use static methods? 

Change either way (adding java8 without generator hooks, or this) is problematic - my vote is that if we are going to claim anything about supporting Java8, we shouldn't break things when they use it.

Patch-set: 2