Update patch set 2
Patch Set 2:

Perhaps we didn't explain it well.
The safe thing is to 'exclude' them; i.e. do not include any static or defender methods when we reply to queries as generators don't know how to handle them.
We believe that is the most compatible behavior and generator will not generate incorrect code when they don't 'see' them.

Patch-set: 2