Update patch set 4
Patch Set 4:

Update: this is the same assertion failure that I mentioned in https://github.com/gwtproject/gwt/issues/9325, but wasn't able to further pin down at the time.

It appears that this is tickled in this case since Spliterator.OfDouble is an interface, yet has a non-abstract method? We seem to be getting to this from AbstractDoubleSpliterator.tryAdvance, which is declared both by implementing Spliterator.OfDouble (which extends OfPrimitive, which extends Spliterator), and also now from extending BaseSpliterator (which implements Spliterator directly as of this change). This mostly from reading local vars in ComputeOverridesAndImplementDefaultMethods.computeOverrides as of the assertion failure.

As of maybeAddSyntheticOverride, it is decided that BaseSpliterator.tryAdvance _is_ abstract (this is true), so createSyntheticAbstractStub is asked to make an abstract stub out of OfDouble.tryAdvance. As a sanity check, it asserts that OfDouble.tryAdvance is abstract, which fails - Spliterator.OfDouble has a non-abstract implementation of this method signature!

I *think* that maybeAddSyntheticOverride's line 289 should also check if interfaceMethod is abstract before asking for the stub to be built. This of course saves the assert from going off, but might also simply make sense for this case. I would be slightly concerned that we would be changing BaseSpliterator rather than just AbstractDoubleSpliterator as we should be done... but I'm pretty far out of my depth in here.

It is very difficult to try any changes in here with so many other tests broken as well (see http://build.gwtproject.org/job/gwt/498/#showFailuresLink). Some of these were broken last night as well, but I havent yet been able to point a finger at a specific place that has caused them to break (though it is very possible that the Collectors patch exposed some of these, despite otherwise compiling and testing correctly).

Patch-set: 4