Update patch set 2
Patch Set 2:

Okay, thanks for reiterating, you are right, I misunderstood the first time. 

To reiterate the suggestion, to make sure I'm arguing the right problem, the proposal is to have the types exposed by the TypeOracle to a given Generator exclude any methods on an interface which were not legal before Java 8: no static methods, and no default methods.

I'm still concerned that this approach can be broken as well, or at least easily made to defy expectations that a developer would have on their code. For example, GWT's Editors UiBinder, and CssResource (within ClientBundle) each permit some mechanism to write in some dsl (editor fields/methods, {path.to.object} or attribute=value, and @eval) code which references other Java types. Other libraries out there do this as well: certainly GXT's XTemplates, and very likely Errai's Templated system as well. 

In these cases, the generator typically needs to resolve the method to invoke so that it knows the return type - UiBinder will call setters or invoke getter (or other no-arg methods?), Editor delegates will call getters/setters on the type they map their editor impl to, etc. If any of those are ever implemented with a default method, or some @eval in a GssResource points to a static method, we'll now be skipping them.

Now of course, this won't be breaking any existing code, or at least I can't see a way to break it just yet. But it certainly will be surprising and confusing to have newly created methods subject to totally different rules than the rest of gwt.

Patch-set: 2