Update patch set 2
Patch Set 2:

(Sorry, was out of range of the internet for the weekend)

I'd be amenable to that, at least in theory, but I'd like to get Daniel's opinion of it too - it seems like the kind of change that requires future commitment of generator fixes, which this general approach was designed to avoid. Of course we aren't saying "no" to contributed changes (as we haven't said "no" over the last year to new features), but a change like this would strongly suggest that we would welcome and encourage changes to bring existing generators kicking and screaming into 2 year old Java.

In other words, I'm saying that this feels like actual, purposeful maintenance, and it has been suggested that we want to avoid giving that impression. 

If everyone is comfortable with this, then I'll abandon this commit, and try for another one. I would ask for a small bit of guidance/suggestion on how to build different oracles based on what the generator calls for, or if each generator already happens to get a completely different oracle already, and so all that is requires is to plumb through this new generator setting to CTUOracleUpdater.

If instead you are suggesting that consensus is to just forbid access to default and static methods and pretend they don't exist, I'd have to disagree at least on the grounds that this makes UiBinder/GssResource unpredictable at best (interfaces extending interfaces, etc), or at worst broken and unmaintained. Unless someone here is jumping on the work of porting either/both of these to APT...?

Patch-set: 2